Numerous updates all round to support sampling at frequencies lower than 1Hz. By support, this is mainly to do with the names of output files. For now, all sampling frequencies are when written out are written out as floats with the decimal point replaced by underscores.

+++ resistics/ioHandlers/transferFunctionReader.py
@@ -138,7 +138,7 @@ class TransferFunctionReader(IOHandler):
         """Format  header values for the internal transfer function file format"""
 
         if "sampleFreq" in self.headers:
-            self.headers["sampleFreq"] = int(self.headers["sampleFreq"])
+            self.headers["sampleFreq"] = float(self.headers["sampleFreq"])
         test = ["insite", "outsite", "remotesite"]
         for header in test:
             if header in self.headers and self.headers[header] == "False":
